+++ src/main/java/JOO/jooshop/payment/model/PaymentRequestDto.java
@@ -3,6 +3,7 @@ package JOO.jooshop.payment.model;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,10 +11,10 @@ import java.util.List;
 public class PaymentRequestDto {
     private Long memberId;
     private Long orderId;
-    private Long price;
+    private BigDecimal price;
     private List<Long> inventoryIdList;
 
-    public PaymentRequestDto(Long memberId, Long orderId, Long price, List<Long> inventoryIdList) {
+    public PaymentRequestDto(Long memberId, Long orderId, BigDecimal price, List<Long> inventoryIdList) {
         this.memberId = memberId;
         this.orderId = orderId;
         this.price = price;
